==> 1) Data Preperation.ipynb <==
# Data Preparation & Exploration

In this notebook, I load, clean and join relevant datafiles, before saving the final product for further analysis

The steps that are performed in this notebook are the following:   
   
1) Data pre-processing   
    - Loading packages   
    - Reading in the data and adding variables where necessary   
    - Selecting only political online articles   
    - Joining the two dataframes   
    - Removing duplicates    
    - Transforming column types    
    - Cleaning up the text data   
    - Upating article length   
    - Removing articles that could not be scraped due to them being behind a paywall   
       
2) NLP to stem text and remove stopwords  

3) Creating dummy variables

4) Saving the clean data

5) Exploring the distribution of the data among different modalities and outlets

### 1) Data pre-processing   
   
#### a) Loading packages

#import pandas
import pandas as pd
from pandas import read_excel
#import numpy
import numpy as np
#load SpaCy
import spacy
#import German language model
import de_core_news_md
#define nlp pipe
nlp = de_core_news_md.load()

#### b) Reading in the data and adding variables where necessary

#read print data
printdata = pd.read_csv("printdata_19Apr_to_08June.csv", encoding = "ISO-8859-1")
#transform data into a dataframe for easier analysis
df_print = pd.DataFrame(printdata)
#add relevant columns for later joining
df_print["Modality"] = "print"
df_print["Teaser"] = ""
df_print["url"] = "none"
#adapt the length column by applying a function
#definingthe function
def removewords(string):
    return string.split(" ")[0].strip()
#applying the function
df_print["Length"] = df_print["Length"].apply(removewords)
#rename columns for later joining
df_print = df_print.rename(columns={"Section":"Category"})
#select relevant columns
df_print = df_print[["ID", "Newspaper", "Date", "Length", "Category", "Author", "Headline", "Teaser", "Article", "Modality", "url"]] 
df_print.head(3)

df_print.groupby("Newspaper").count()

len(df_print)

#read online data
df_online = read_excel("online_data_complete.xlsx")
#rename columns
df_online = df_online.rename(columns={"_source.title_rss": "Headline", "_source.doctype": "Newspaper", "_source.publication_date": "Date", "_source.teaser_rss": "Teaser", "_source.text": "Article", "_source.category": "Category", "_source.url":"url"})
#add columns for later joining
df_online["Length"] = 0 
df_online["Modality"] = "online"
#reset index and add unique ID column
df_onine = df_online.reset_index(drop=True)
df_online["ID"] = range(100000, 100000+len(df_online))
df_online["Author"] = ""
#select relevant columns
df_online = df_online[["ID", "Newspaper", "Date", "Length", "Category", "Author", "Headline", "Teaser", "Article", "Modality", "url"]]
df_online.head(3)

len(df_online)

df_online.groupby("Newspaper").count()

#### c) Selecting only political online articles

len(df_online)

#first I remove articles that weren't parsed correctly
df_online = df_online[~df_online["Article"].str.startswith('    ', na=False)]
#next I remove Impressum articles from die Süddeutsche
df_online = df_online[~df_online["Article"].str.contains("HERAUSGEGEBEN VOM SÜDDEUTSCHEN VERLAG", na= False)]
#after that I do the same for Stuttgarter Zeitung
df_online = df_online[~df_online["Article"].str.contains("Impressum Stuttgarter Zeitung Verlagsgesellschaft", na= False)]
#next I include only political articles for the online dataset
relevant_articles = df_online[df_online['url'].astype(str).str.contains("politik")]
relevant_articles2 = df_online[df_online["Category"].astype(str).str.contains("Politik")]
df_online = pd.concat([relevant_articles, relevant_articles2])
len(df_online)

#### d) Joining the two dataframes

#join the two datasets
df = pd.concat([df_online, df_print])
len(df)

#inspect df
df.head()

#### e) Remove duplicates

#check the size of the dataset
len(df)

#create a list of duplicates
duplicates = df[df.duplicated(["Article"])]
len(duplicates)

#remove duplicates
df.drop_duplicates(subset ="Article", keep = "first", inplace = True) 
len(df)

#check for duplicates
duplicates = df[df.duplicated(["Article"])]
len(duplicates)

#### f) Transform column types 

df.dtypes

df["Article"] = df["Article"].astype(str)
df["Length"] = df["Length"].astype(int)
df["Category"] = df["Category"].astype(str)
df["Author"] = df["Author"].astype(str)
df["Modality"] = df["Modality"].astype(str)
#see if it worked
df.dtypes

#### g) Cleaning up the text data

#show the 1000 most frequent words in the df, in order to spot potential issues that require cleaning
from collections import Counter
Counter(" ".join(df["Article"]).split()).most_common(1000)

Result: There are several formatting errors. As a result, strings like "\xa96" are very common.

#import ftfy to fix formatting errors
import ftfy

#define function to clean the text with 
def fixtext(string):
    return ftfy.fix_text(string)

#apply function
df["Article"] = df["Article"].apply(fixtext)

#check if it worked 
df[df['Article'].str.contains("\xa96")]

Result: It seems like it have worked. I tried this not only with this, but also with some other strings.   
There are however still some other issues that I noticed while manually exploring some of the articles (not included in this Notebook for simplicity). For example, Articles from "Der Tagesspiegel" often include a graphic and its title at the end. Issues like this are addressed in the following.

#remove everything after and including "Graphic" -> removes pictures and their headings from Tagesspiegel articles
def removegraphic(string):
    return string.split("Graphic")[0].strip()
#explore how pervailing this issue is
print(len(df[df['Article'].str.contains("Graphic")]))

df["Article"] = df["Article"].apply(removegraphic)

#see if the code worked
print(len(df[df['Article'].str.contains("Graphic")]))

There are also some more formatting errors that ftfy did not capture. They are addressed in the following.

#removes the \xa0 string from the text
def reformat(string):
    return string.replace("\xa0", " ")

#see how prevailing this issue is
print(len(df[df['Article'].str.contains("\xa0")]))

#apply the function
df["Article"] = df["Article"].apply(reformat)

#see if it worked
print(len(df[df['Article'].str.contains("\xa0")]))

#removes the \xa0 string from the text
def reformat2(string):
    return string.replace("\xad", " ")

#see how prevailing this issue is
print(len(df[df['Article'].str.contains("\xad")]))

#apply the function
df["Article"] = df["Article"].apply(reformat2)

#see if it worked
print(len(df[df['Article'].str.contains("\xad")]))

#removes all instances of the \' string from the text. This string denotes the German genitive endinng
#This ending is not important for any of the following analysis an might impede readability metrics
def reformat3(string):
    return string.replace("\'", " ")

#see how prevailing this issue is
print(len(df[df['Article'].str.contains("\'")]))

#apply the function
df["Article"] = df["Article"].apply(reformat3)

#see if it worked
print(len(df[df['Article'].str.contains("\'")]))

#replaces three spaces with just one
def reformat4(string):
    return string.replace("   ", " ")

#see how prevailing this issue is
print(len(df[df['Article'].str.contains("   ")]))

#apply the function
df["Article"] = df["Article"].apply(reformat4)

#see if it worked
print(len(df[df['Article'].str.contains("   ")]))

#replaces dpa which is often named as a source at the end of an article
def reformat5(string):
    return string.replace("dpa", "")

#see how prevailing this issue is
print(len(df[df['Article'].str.contains("dpa")]))

#apply the function
df["Article"] = df["Article"].apply(reformat5)

#see if it worked
print(len(df[df['Article'].str.contains("dpa")]))

#replaces fullstops with a fullstop followed by a space in order to avoid confounds of the readability meausure
def reformat6(string):
    return string.replace(".", ". ")

#apply the function
df["Article"] = df["Article"].apply(reformat6)

#replace double spaces with single spaces - this way this operation only affected the cases where a fullstop was not followed by a space
def reformat7(string):
    return string.replace("  ", " ")

#apply the function
df["Article"] = df["Article"].apply(reformat7)

df.head(5)

df["Article"][6]

#### h) Update article length

#function that counts the number of words in articles
def count_words(text):
    doc = nlp(text)
    tokens = [token.text for token in doc]
    tokens_no_punctuation = []
    punctuation = [",", ".", ":", "!", "?", "'", ",", "  ", ")", "(", "-", '"', "   ", ";", "»", "«", "'", "/"]
    for w in tokens:
        if w not in punctuation:
            tokens_no_punctuation.append(w)
    return len(tokens_no_punctuation)

#apply function to df
df["Length"] = [count_words(text) for text in df["Article"].astype(str)]
                   
#inspect data
df.head(3)

#### i) Remove articles that could not be scraped due to them being behind a paywall

#check how many articles have fewer than 40 words.
len(df[df["Length"] <40])

#remove the short articles
df = df[~(df["Length"] <=40)]
#check if the code worked
print(len(df), len(df[df["Length"] <40]))

### 2) NLP to stem text and remove stopwords

from nltk.corpus import stopwords 
stop_words = set(stopwords.words('german')) 
from nltk.stem.snowball import SnowballStemmer
stemmer = SnowballStemmer("german")
from collections import Counter

#function to remove stopwords
def remove_stopwords_and_stem(text):
    #apply SpaCy to tokenise
    doc = nlp(text)
    #create a list of tokens
    tokens = [token.text for token in doc]
    #create an empty list for all words in a text that are not stopwords
    no_stopwords = []
    #loop over the tokens list and append non-stopwords to the no_stopwords list
    for w in tokens: 
        if w not in stop_words: 
            no_stopwords.append(w)
    #stem all words in the list
    stems=""
    for word in no_stopwords:
        stems=stems + stemmer.stem(word) + " "
    #return the final stems as a single string
    return(stems)

#apply functions in order to...

#... create a column with the cleaned article text (where stopwords are removed and all words are stemmed)
df["clean text"] = [remove_stopwords_and_stem(text) for text in df["Article"]]

#... create a column with the number of words for the cleaned text
df["words in clean text"] = [count_words(text) for text in df["clean text"]]

#inspect data
df.head(3)

### 3) Creating dummy variables

#### a) for reach

regional = ["Aachener Zeitung", "Rheinische", "Stuttgarter", "aachener", "rheinische, stuttgarter"]
national = ["Tagesspiegel", "tagesspiegel", "Welt", "welt", "Süddeutsche", "süddeutschet"]

def create_reach_dummy(row):
        if row == "Aachener Zeitung" or row == "aachener zeitung (www)" or row == "Stuttgarter Zeitung" or row == "stuttgarter zeitung (www)" or row == "Rheinische Post" or row == "rheinische post (www)":
            return 0
        else:
            return 1

df["reach_dummy"] = [create_reach_dummy(row) for row in df["Newspaper"]]

#### b) for modality

cond1 = df.Modality.str.contains("print")
cond2 = df.Modality.str.contains("online")
df["modality_dummy"] = np.where(cond1, 1, np.where(cond2, 0, 2) )

### 4) Saving the clean and complete datafile for further analysis

df.to_excel("complete_data_cleaned.xlsx")

### 5) Exploring the data distribution

df.groupby("Newspaper").count()

len(df)
