==> 3) Assessing impartiality.ipynb <==
# Assessing impartiality
   
In this notebook, I use three classifiers to predict the impartiality indicators neutrality, balance of actors, and balance of viewpoints, before combining them into an overall index. 

## Load packages

#import relevant packages
import pandas as pd
from pandas import read_excel
import joblib
import numpy as np
from sklearn import svm
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import LinearSVC
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.pipeline import Pipeline
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score
from sklearn.model_selection import RandomizedSearchCV, ShuffleSplit, GridSearchCV, train_test_split
from sklearn.feature_selection import SelectKBest, chi2
from sklearn.linear_model import SGDClassifier
from sklearn.naive_bayes import GaussianNB, MultinomialNB

## Import classifiers

classifier_neu = joblib.load("classifier_neu.pkl")
classifier_boa = joblib.load("classifier_boa.pkl")
classifier_bov = joblib.load("classifier_bov.pkl")

## Read data

#read in the complete cleaned dataset
df = read_excel("complete_data_cleaned.xlsx")
mca = read_excel("mca_cleaned.xlsx")
#inspect data
print(len(df), len(mca))

## Vectorise data

# create a count vectorizer object 
count_vect = CountVectorizer(analyzer="word", 
                             token_pattern=r"\w{1,10}", 
                             min_df = 10, 
                             max_df = 1., 
                             max_features=200)

#features for training
count_vectors_clean = count_vect.fit_transform(df["clean text"]).toarray()
count_vectors_full = count_vect.fit_transform(df["Article"]).toarray()

#features for testing
count_vectors_clean2 = count_vect.fit_transform(mca["clean text"]).toarray()
count_vectors_full2 = count_vect.fit_transform(mca["Article"]).toarray()

#inspect the shape
print(count_vectors_clean.shape)
print(count_vectors_clean2.shape)

# unigrams
tfidf_vect_ug = TfidfVectorizer(analyzer='word', 
                             token_pattern=r'\w{1,}',
                             ngram_range = (1,1),
                             min_df = 10, 
                             max_df = 1., 
                             max_features=200)

# bigrams
tfidf_vect_bg = TfidfVectorizer(analyzer='word', 
                             token_pattern=r'\w{1,}',
                             ngram_range = (2,2),
                             min_df = 10, 
                             max_df = 1., 
                             max_features=5) #otherwise there is an error

# unigrams and bigrams
tfidf_vect_ubg = TfidfVectorizer(analyzer='word', 
                             token_pattern=r'\w{1,}',
                             ngram_range = (1,2),
                             min_df = 10, 
                             max_df = 1., 
                             max_features=200)


#unigrams
tfidf_ug_vectors_clean = tfidf_vect_ug.fit_transform(df["clean text"]).toarray()
tfidf_ug_vectors_clean2 = tfidf_vect_ug.fit_transform(mca["clean text"]).toarray()
tfidf_ug_vectors_full = tfidf_vect_ug.fit_transform(df["Article"]).toarray()
tfidf_ug_vectors_full2 = tfidf_vect_ug.fit_transform(mca["Article"]).toarray()
#bigrams
tfidf_bg_vectors_clean = tfidf_vect_bg.fit_transform(df["clean text"]).toarray()
tfidf_bg_vectors_clean2 = tfidf_vect_bg.fit_transform(mca["clean text"]).toarray()
tfidf_bg_vectors_full = tfidf_vect_bg.fit_transform(df["Article"]).toarray()
tfidf_bg_vectors_full2 = tfidf_vect_bg.fit_transform(mca["Article"]).toarray()
#uni and bigrams
tfidf_ubg_vectors_clean = tfidf_vect_ubg.fit_transform(df["clean text"]).toarray()
tfidf_ubg_vectors_clean2 = tfidf_vect_ubg.fit_transform(mca["clean text"]).toarray()
tfidf_ubg_vectors_full = tfidf_vect_ubg.fit_transform(df["Article"]).toarray()
tfidf_ubg_vectors_full2 = tfidf_vect_ubg.fit_transform(mca["Article"]).toarray()

#inspect the shape
print(tfidf_ug_vectors_clean.shape)
print(tfidf_ug_vectors_clean2.shape)

## Apply the classifiers

results = classifier_neu.predict(count_vectors_clean)
#save the results in a new column
df["neutrality dummy"] = results
#inspect the results
df["neutrality dummy"].value_counts()

results = classifier_boa.predict(tfidf_ubg_vectors_full)
#save the results in a new column
df["balance of actors dummy"] = results
#inspect the results
df["balance of actors dummy"].value_counts()

results = classifier_boa.predict(count_vectors_full)
#save the results in a new column
df["balance of viewpoints dummy"] = results
#inspect the results
df["balance of viewpoints dummy"].value_counts()

## Create index

df["impartiality"] = df["balance of viewpoints dummy"] + df["balance of actors dummy"] + df["neutrality dummy"]

## Inspect data

df.groupby("Newspaper").mean()

## Save data

df.to_excel("complete_data_cleaned_with_impartiality.xlsx")
