==> 4) Assessing emotionality.ipynb <==
# Emotionality & sentiment assessment

## Assessing emotionality with SpaCy & NLTK

In this Jupyter notebook, I assess the emotionality of each article, as well as its negativity and positivity (I also included sentiment functions but ended up not using them).
I do so using two different sentiment dictionaries, in order to have a comparison of sentiment measurements.   
The first sentiment dictionary is taken from the SentiWS corpus. The second is taken from the Rauh sentiment dictionary and, unlike the SentiWS dicitonary, accounts for negations (to an extent).

The steps that are performed in this notebook are the following:   
   
1) Loading necessary packages
   
2) Loading the cleaned data
   
3) Assessing emotionality with the SentiWS corpus   
    - Creating lists of positive and negative words   
    - Creating functions to count positve and negative words   
    - Applying counting functions to get the overall number as well as the ratio of positive and negative words   
    - Calculating overall emotionality   
    
4) Computing overall article sentiment according to the SentiWS corpus    
    - Creating a sentiment function   
    - Applying the function to the dataframe   
    - Inspecting the data   
   
5) Assessing emotionality with the Rauh sentiment dictionary     
    - Creating lists of positive and negative words   
    - Creating functions to count positve and negative words   
    - Applying counting functions to get the overall number as well as the ratio of positive and negative words   
    - Calculating overall emotionality   
    
6) Computing overall article sentiment according to Rauh sentiment dictionary   
    - Creating a sentiment function that accounts for negations   
    - Applying the function to the dataframe   
    - Inspecting the data   
    
7) Saving the data with the newly created variables

### 1) Loading necessary packages

#import pandas
import pandas as pd
from pandas import read_excel
#import numpy
import numpy as np
#load SpaCy
import spacy
#import German language model
import de_core_news_md
#define nlp pipe
nlp = de_core_news_md.load()

### 2) Loading the cleaned data

df = read_excel("complete_data_cleaned.xlsx")

### 3) Assessing emotionality with the SentiWS corpus

#### a) Creating lists of positive and negative words

#read in the SentiWS text file with positive words
pos = pd.read_csv("SentiWS_v1.8c_Positive.txt",sep='\t', names=["word", "value", "forms"])
#clean the text file
pos[['word','wordtype']] = pos.word.str.split("|",expand=True) 
#create list of positive words (empty for now)
pos_list = []
#loop over the df in order to create a list with all positive words
for e in pos["word"]:
    pos_list.append(e)

#same as above
neg = pd.read_csv("SentiWS_v1.8c_Negative.txt",sep='\t', names=["word", "value", "forms"])
neg[['word','wordtype']] = neg.word.str.split("|",expand=True) 
neg_list = []
for e in neg["word"]:
    neg_list.append(e)

#### b) Creating functions to count positve and negative words

#function to detect positive words
def count_positivewords(text):
    doc = nlp(text)
    tokens = [token.text for token in doc]
    poscount = 0
    for w in tokens:
        if w in pos_list:
            poscount+=1
    return poscount

#function to detect negative words
def count_negativewords(text):
    doc = nlp(text)
    tokens = [token.text for token in doc]
    negcount = 0
    for w in tokens:
        if w in neg_list:
            negcount+=1
    return negcount

#### c) Applying counting functions to get the overall number as well as the ratio of positive and negative words

#apply functions in order to...

#... create a column with the number of positive words
df["positive words"] = [count_positivewords(text) for text in df["clean text"]]

#... create a column with the number of negative words
df["negative words"] = [count_negativewords(text) for text in df["clean text"]]

#creating a new column for overall emotionality (percenage of overall words that are negative)
df["negativity ratio"] = df["negative words"]/df["words in clean text"]
#creating a new column for overall emotionality (percenage of overall words that are positve)
df["positivity ratio"] = df["positive words"]/df["words in clean text"]

#### d) Use the number of positive and negative words per article to calculate overall emotionality

#creating a new column for overall emotionality 
df["emotionality"] = (df["negative words"]+df["positive words"])
#creating a new column for overall emotionality ratio (percenage of overall words that are positve or negative)
df["emotionality ratio"] = (df["negative words"]+df["positive words"])/df["words in clean text"]

#inspect data
df.head(3)

### 4) Computing overall article sentiment according to the SentiWS corpus
    
#### a) Creating a sentiment function

def compute_sentiment(text):
    doc = nlp(text)
    tokens = [token.text for token in doc]
    word_list = []
    for e in tokens:
        if e in pos_list:
            word_list.append(e)
        elif e in neg_list:
            word_list.append(e)
    df1 = pos[pos['word'].isin(word_list)]
    df2 = neg[neg['word'].isin(word_list)]
    dfs = [df1, df2]
    df_final = pd.concat(dfs)
    sentimentscore = df_final["value"].sum()
    return sentimentscore

#### b) Applying the function to the dataframe

#first a function for the overall sentiment score
df["sentimentscore"] = [compute_sentiment(text) for text in df["clean text"]]
#then calculate a ratio to account for article length
df["sentimentscore ratio"] = df["sentimentscore"]/df["words in clean text"]

#### c) Data overview

#overview of how many articles are in the dataset
df.groupby("Newspaper").count()

#overview of average values for the newly created columns
df.groupby("Newspaper").mean()

### 5) Assessing emotionality with the Rauh sentiment dictionary

#read in the rauh dictionary
rauh = pd.read_csv("rauh_sentiment.csv")
#remove whitespaces
rauh["feature"] = rauh["feature"].str.strip()
#inspect the dictionary
rauh.head(3)

#split the dictionary into a positive and a negative one
pos_rauh = rauh.loc[rauh["sentiment"] == 1]
neg_rauh = rauh.loc[rauh["sentiment"] == -1]
#delete spaces at the end of each string


#create a list of positive words
pos_list_rauh = []
for e in pos_rauh["feature"]:
    pos_list_rauh.append(e)
#and a list of negative words
neg_list_rauh = []
for e in neg_rauh["feature"]:
    neg_list_rauh.append(e)

#function to detect positive words
def count_positivewords_rauh(text):
    doc = nlp(text)
    tokens = [token.text for token in doc]
    poscount = 0
    for w in tokens:
        if w in pos_list_rauh:
            poscount+=1
    return poscount

#function to detect negative words
def count_negativewords_rauh(text):
    doc = nlp(text)
    tokens = [token.text for token in doc]
    negcount = 0
    for w in tokens:
        if w in neg_list_rauh:
            negcount+=1
    return negcount

#... create a column with the number of positive words
df["positive words rauh"] = [count_positivewords_rauh(text) for text in df["clean text"]]

#... create a column with the number of negative words
df["negative words rauh"] = [count_negativewords_rauh(text) for text in df["clean text"]]

#inspect the df
df.head(3)

#creating a new column for overall emotionality 
df["emotionality  rauh"] = (df["negative words rauh"]+df["positive words rauh"])
#creating a new column for overall emotionality ratio (percenage of overall words that are positve or negative)
df["emotionality ratio rauh"] = (df["negative words rauh"]+df["positive words rauh"])/df["words in clean text"]
#creating a new column for overall emotionality (percenage of overall words that are negative)
df["negativity ratio rauh"] = df["negative words rauh"]/df["words in clean text"]
#creating a new column for overall emotionality (percenage of overall words that are positve)
df["positivity ratio rauh"] = df["positive words rauh"]/df["words in clean text"]

### 6) Computing overall sentiment with the Rauh sentiment dictionary   
   
#### a) Writing a function to compute sentiment

#Rauh designed an extra dictionary that also includes negations. 
#However, those negations are always the same. 
#I incorporated Rauh's approach towards accounting for negations by creating a seperate negation list and computing sentiment accordingly.

#create a list of negations
negations = ["nicht","nichts","kein","keine","keinen"]
#remove negations from negative list
neg_rauh = neg_rauh[~neg_rauh["feature"].isin(negations)]

def compute_sentiment_rauh(text):
    #tokenise the text
    doc = nlp(text)
    #create a list of sentences
    sentences = [sent.text for sent in doc.sents]
    #create list of words
    wordlist = []
    wordlist_new2 = []
    wordlist_new = []
    #create lists for negations and their sentiment value
    negationlist = []
    sentimentscore_negations = []
    #analyse a document sentence by sentence
    for sentence in sentences:
        #tokenise the sentence
        sentence = nlp(sentence)
        #save tokens as a list
        tokens = [token.text for token in sentence]
        #append the tokens to a list of words
        wordlist.append(tokens[0])
        #check for negations
        for last_item, e in zip(tokens, tokens[1:]):
            if last_item in negations:
                negationlist.append(e)
            else:
                wordlist.append(e)
        #append positive an negative words to a new list
        for e in wordlist:
            if e in pos_list_rauh:
                wordlist_new.append(e)
            elif e in neg_list_rauh:
                wordlist_new.append(e)
    #analyse negationlist and save a list of the words' sentiment value
    for e in negationlist:
        if e in pos_list_rauh:
            sentimentscore_negations.append(-1)
        elif e in neg_list_rauh:
            sentimentscore_negations.append(1)
    #create and merge dataframes
    df1 = pos_rauh[pos_rauh['feature'].isin(wordlist_new)]
    df2 = neg_rauh[neg_rauh['feature'].isin(wordlist_new)]
    dfs = [df1, df2]
    df_final = pd.concat(dfs)
    #compute sentimentscore without negations
    sentimentscore = df_final["sentiment"].sum()
    #compute sentimentscore for only the negations
    sentimentscore_negations = pd.DataFrame(sentimentscore_negations, columns=["sentiment"])
    sentimentscore_negations = sentimentscore_negations["sentiment"].sum()
    #add the two sentiment scores together for a final sentiment value
    sentimentscore_final = sentimentscore + sentimentscore_negations
    return sentimentscore_final

#### b) Applying the function to calculate both overall sentiment and sentiment ratio

#compute the sentiment
df["sentimentscore_rauh"] = [compute_sentiment_rauh(text) for text in df["clean text"]]
#calcuate the sentiment ratio to account for length
df["sentimentscore_rauh_ratio"] = df["sentimentscore_rauh"]/df["words in clean text"]

#### c) Data overview

#inspect data
df.groupby("Newspaper").mean()

## 7) Saving the data

df.to_excel("complete_data_cleaned_with_emotionality.xlsx")
