==> 5) Assessing readability.ipynb <==
## Assessing an article's readability score

In this notebook, I assess an individual article's readability score. I do so with help of the textstat package.

### 1) Loading necessary packages and data

#import pandas
import pandas as pd
from pandas import read_excel

df = read_excel("complete_data_cleaned.xlsx")

#importing the textstat package
import textstat as ts
#set the language to German
ts.set_lang("de")

### 2) Testing the fromula

#Formula: 180 - ASL - 58,5* 
#Thus the following should be 180 - 5 - (58,5*6/5) = 175-70,2 = 104.8
ts.flesch_reading_ease("Das ist ein leichter Satz.")

### 3) Calculating readability

#calculate readability score per row for teaser and text
#note that teaser data is only available for online articles
df["readabilityscore_text"] = df["Article"].apply (ts.flesch_reading_ease)
df.head(3)

### 4) Comparing outlets

#grouping by outlet
df.groupby("Newspaper").mean()

df.head(10)

df["Article"][2]

### 5) Save output

df.to_excel("complete_data_cleaned_with_reliability.xlsx")
