==> 6) Hypothesis testing actor diversity.ipynb <==
# Hypothesis testing: Impartiality

In this Notebook, the impartiality data is explored and the related hypotheses are (partially) tested.

## Loading packages and data

#import the necessary packages
import pandas as pd
from pandas import read_excel
from scipy import stats
from scipy.stats import mannwhitneyu

#read and inspect the data
df = read_excel("complete_data_cleaned_with_impartiality.xlsx")
df.head(3)

df.columns

## Select relevant columns

df = df[["ID", "Newspaper", "Length", "Article", "modality_dummy", "reach_dummy", "clean text", 
         "neutrality dummy", "balance of actors dummy", "balance of viewpoints dummy", "impartiality"]]
len(df)

## Descriptives

print("Overall sample")
print(df["impartiality"].mean())
print(df["impartiality"].std())

print("Print sample")
print(df[df["modality_dummy"] == 1]["impartiality"].mean())
print(df[df["modality_dummy"] == 1]["impartiality"].std())
print("")
print("Online sample")
print(df[df["modality_dummy"] == 0]["impartiality"].mean())
print(df[df["modality_dummy"] == 0]["impartiality"].std())

print("National sample")
print(df[df["reach_dummy"] == 1]["impartiality"].mean())
print(df[df["reach_dummy"] == 1]["impartiality"].std())
print("")
print("Regional sample")
print(df[df["reach_dummy"] == 0]["impartiality"].mean())
print(df[df["reach_dummy"] == 0]["impartiality"].std())

## Hypothesis test

print("RQ3a: Reach")
print(mannwhitneyu(df[df["reach_dummy"] == 1]["impartiality"], df[df["reach_dummy"] == 0]["impartiality"]))

print("RQ3b: Modality")
print(mannwhitneyu(df[df["modality_dummy"] == 1]["impartiality"], df[df["modality_dummy"] == 0]["impartiality"]))

## Comparisons

df[["Newspaper", "neutrality dummy", "balance of actors dummy", "balance of viewpoints dummy", "impartiality"]].groupby("Newspaper").mean()

df[["Newspaper", "neutrality dummy", "balance of actors dummy", "balance of viewpoints dummy", "impartiality"]].groupby("Newspaper").std()

## Means and standard deviations

print("AZ")
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["impartiality"].mean())
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["impartiality"].std())
print("RP")
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["impartiality"].mean())
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["impartiality"].std())
print("STZ")
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["impartiality"].mean())
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["impartiality"].std())
print("Welt")
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["impartiality"].mean())
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["impartiality"].std())
print("TS")
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["impartiality"].mean())
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["impartiality"].std())
print("SZ")
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["impartiality"].mean())
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["impartiality"].std())
