==> 7) Hypothesis testing impartiality.ipynb <==
# Hypothesis testing: readability

In this Notebook, the readability data is explored and the related hypotheses are (partially) tested.

## Loading packages and data

#import the necessary packages
import pandas as pd
from pandas import read_excel
from scipy import stats
from scipy.stats import mannwhitneyu

#read and inspect the data
df = read_excel("complete_data_cleaned_with_actor_diversity.xlsx")
df.head(3)

df.columns

## Select relevant columns

df = df[["ID", "Newspaper", "Length", "Article", "modality_dummy", "reach_dummy", "clean text", 
         "persons", "organisations", "national_elite_actors", "national_opposition_actors", "regional_elite_actors", 
         "regional_opposition_actors", "person dummy", "organisation dummy","nea dummy", "noa dummy", "rea dummy",
         "roa dummy", "diversity index", "ea dummy", "oa dummy", "diversity index all actors"]]
len(df)

## Descriptives

print("Overall sample")
print(df["diversity index"].mean())
print(df["diversity index"].std())

print("Print sample")
print(df[df["modality_dummy"] == 1]["diversity index"].mean())
print(df[df["modality_dummy"] == 1]["diversity index"].std())
print("")
print("Online sample")
print(df[df["modality_dummy"] == 0]["diversity index"].mean())
print(df[df["modality_dummy"] == 0]["diversity index"].std())

print("National sample")
print(df[df["reach_dummy"] == 1]["diversity index"].mean())
print(df[df["reach_dummy"] == 1]["diversity index"].std())
print("")
print("Regional sample")
print(df[df["reach_dummy"] == 0]["diversity index"].mean())
print(df[df["reach_dummy"] == 0]["diversity index"].std())

## Hypothesis test

print("RQ3a: Reach")
print(mannwhitneyu(df[df["reach_dummy"] == 1]["diversity index"], df[df["reach_dummy"] == 0]["diversity index"]))

print("RQ3b: Moality")
print(mannwhitneyu(df[df["modality_dummy"] == 1]["diversity index"], df[df["modality_dummy"] == 0]["diversity index"]))

## Comparisons

#grouping by outlet
df[["Newspaper", "diversity index","Length"]].groupby("Newspaper").mean()

df[["Newspaper", "diversity index","Length"]].groupby("Newspaper").std()

## Means and standard deviations

df[df["modality_dummy"] == 0]["Length"].mean()

df[df["modality_dummy"] == 1]["Length"].mean()

df[df["reach_dummy"] == 0]["Length"].mean()

df[df["reach_dummy"] == 1]["Length"].mean()

df["Length"].mean()

df["Length"].std()

df[df["modality_dummy"] == 0]["Length"].std()

df[df["modality_dummy"] == 1]["Length"].std()

df[df["reach_dummy"] == 0]["Length"].std()

df[df["reach_dummy"] == 1]["Length"].std()

print("AZ")
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["Length"].mean())
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["Length"].std())
print("RP")
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["Length"].mean())
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["Length"].std())
print("STZ")
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["Length"].mean())
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["Length"].std())
print("Welt")
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["Length"].mean())
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["Length"].std())
print("TS")
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["Length"].mean())
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["Length"].std())
print("SZ")
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["Length"].mean())
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["Length"].std())

print("AZ")
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["diversity index"].mean())
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["diversity index"].std())
print("RP")
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["diversity index"].mean())
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["diversity index"].std())
print("STZ")
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["diversity index"].mean())
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["diversity index"].std())
print("Welt")
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["diversity index"].mean())
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["diversity index"].std())
print("TS")
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["diversity index"].mean())
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["diversity index"].std())
print("SZ")
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["diversity index"].mean())
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["diversity index"].std())
