==> 8) Hypothesis testing emotionality.ipynb <==
# Hypothesis testing: readability

In this Notebook, the readability data is explored and the related hypotheses are tested.

## Loading packages and data

#import the necessary packages
import pandas as pd
from pandas import read_excel
from scipy import stats
from scipy.stats import mannwhitneyu

#read and inspect the data
df = read_excel("complete_data_cleaned_with_emotionality.xlsx")
df.head(3)

## Select relevant columns

df = df[["ID", "Newspaper", "Length", "Article", "modality_dummy", "reach_dummy", "clean text", "emotionality ratio", "emotionality ratio rauh", "negativity ratio", "negativity ratio rauh"]]
len(df)

## Descriptives

print("Overall sample - Emotionality")
print("SentiWS")
print(df["emotionality ratio"].mean())
print(df["emotionality ratio"].std())
print("Rauh sentiment dictionary")
print(df["emotionality ratio rauh"].mean())
print(df["emotionality ratio rauh"].std())

print("Overall sample - Negativity")
print("SentiWS")
print(df["negativity ratio"].mean())
print(df["negativity ratio"].std())
print("Rauh sentiment dictionary")
print(df["negativity ratio rauh"].mean())
print(df["negativity ratio rauh"].std())

print("Senti WS")
print("Print sample - Emotionality")
print(df[df["modality_dummy"] == 1]["emotionality ratio"].mean())
print(df[df["modality_dummy"] == 1]["emotionality ratio"].std())
print("")
print("Online sample - Emotionality")
print(df[df["modality_dummy"] == 0]["emotionality ratio"].mean())
print(df[df["modality_dummy"] == 0]["emotionality ratio"].std())
print("")
print("")
print("")
print("Rauh sentiment dictionary")
print("Print sample - Emotionality")
print(df[df["modality_dummy"] == 1]["emotionality ratio rauh"].mean())
print(df[df["modality_dummy"] == 1]["emotionality ratio rauh"].std())
print("")
print("Online sample - Emotionality")
print(df[df["modality_dummy"] == 0]["emotionality ratio rauh"].mean())
print(df[df["modality_dummy"] == 0]["emotionality ratio rauh"].std())

print("Senti WS")
print("Print sample - Negativity")
print(df[df["modality_dummy"] == 1]["negativity ratio"].mean())
print(df[df["modality_dummy"] == 1]["negativity ratio"].std())
print("")
print("Online sample - Negativity")
print(df[df["modality_dummy"] == 0]["negativity ratio"].mean())
print(df[df["modality_dummy"] == 0]["negativity ratio"].std())
print("")
print("")
print("")
print("Rauh sentiment dictionary")
print("Print sample - Negativity")
print(df[df["modality_dummy"] == 1]["negativity ratio rauh"].mean())
print(df[df["modality_dummy"] == 1]["negativity ratio rauh"].std())
print("")
print("Online sample - Negativity")
print(df[df["modality_dummy"] == 0]["negativity ratio rauh"].mean())
print(df[df["modality_dummy"] == 0]["negativity ratio rauh"].std())

print("Senti WS")
print("National sample - Emotionality")
print(df[df["reach_dummy"] == 1]["emotionality ratio"].mean())
print(df[df["reach_dummy"] == 1]["emotionality ratio"].std())
print("")
print("Regional sample - Emotionality")
print(df[df["reach_dummy"] == 0]["emotionality ratio"].mean())
print(df[df["reach_dummy"] == 0]["emotionality ratio"].std())
print("")
print("")
print("")
print("Rauh sentiment dictionary - Emotionality")
print("National sample")
print(df[df["reach_dummy"] == 1]["emotionality ratio rauh"].mean())
print(df[df["reach_dummy"] == 1]["emotionality ratio rauh"].std())
print("")
print("Regional sample - Emotionality")
print(df[df["reach_dummy"] == 0]["emotionality ratio rauh"].mean())
print(df[df["reach_dummy"] == 0]["emotionality ratio rauh"].std())

print("Senti WS")
print("National sample - Negativity")
print(df[df["reach_dummy"] == 1]["negativity ratio"].mean())
print(df[df["reach_dummy"] == 1]["negativity ratio"].std())
print("")
print("Regional sample - Negativity")
print(df[df["reach_dummy"] == 0]["negativity ratio"].mean())
print(df[df["reach_dummy"] == 0]["negativity ratio"].std())
print("")
print("")
print("")
print("Rauh sentiment dictionary")
print("National sample - Negativity")
print(df[df["reach_dummy"] == 1]["negativity ratio rauh"].mean())
print(df[df["reach_dummy"] == 1]["negativity ratio rauh"].std())
print("")
print("Regional sample - Negativity")
print(df[df["reach_dummy"] == 0]["negativity ratio rauh"].mean())
print(df[df["reach_dummy"] == 0]["negativity ratio rauh"].std())

## Hypothesis test

print("RQ3a: Reach")
print("SW")
print(mannwhitneyu(df[df["reach_dummy"] == 1]["emotionality ratio"], df[df["reach_dummy"] == 0]["emotionality ratio"]))
print("Rauh")
print(mannwhitneyu(df[df["reach_dummy"] == 1]["emotionality ratio rauh"], df[df["reach_dummy"] == 0]["emotionality ratio rauh"]))

print("RQ3b: Modality")
print("SW")
print(mannwhitneyu(df[df["modality_dummy"] == 1]["emotionality ratio"], df[df["modality_dummy"] == 0]["emotionality ratio"]))
print("Rauh")
print(mannwhitneyu(df[df["modality_dummy"] == 1]["emotionality ratio rauh"], df[df["reach_dummy"] == 0]["emotionality ratio rauh"]))

## Comparisons

#grouping by outlet
df.groupby("Newspaper").mean()

#grouping by outlet
df.groupby("Newspaper").std()

print("AZ")
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["emotionality ratio rauh"].mean())
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["emotionality ratio rauh"].std())
print("RP")
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["emotionality ratio rauh"].mean())
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["emotionality ratio rauh"].std())
print("STZ")
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["emotionality ratio rauh"].mean())
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["emotionality ratio rauh"].std())
print("Welt")
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["emotionality ratio rauh"].mean())
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["emotionality ratio rauh"].std())
print("TS")
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["emotionality ratio rauh"].mean())
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["emotionality ratio rauh"].std())
print("SZ")
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["emotionality ratio rauh"].mean())
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["emotionality ratio rauh"].std())

print("AZ")
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["negativity ratio rauh"].mean())
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["negativity ratio rauh"].std())
print("RP")
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["negativity ratio rauh"].mean())
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["negativity ratio rauh"].std())
print("STZ")
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["negativity ratio rauh"].mean())
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["negativity ratio rauh"].std())
print("Welt")
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["negativity ratio rauh"].mean())
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["negativity ratio rauh"].std())
print("TS")
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["negativity ratio rauh"].mean())
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["negativity ratio rauh"].std())
print("SZ")
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["negativity ratio rauh"].mean())
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["negativity ratio rauh"].std())
