==> 9) Hypothesis testing readability.ipynb <==
# Hypothesis testing: readability

In this Notebook, the readability data is explored and the related hypotheses are tested.

## Loading packages and data

#import the necessary packages
import pandas as pd
from pandas import read_excel
from scipy import stats
from scipy.stats import mannwhitneyu

#read and inspect the data
df = read_excel("complete_data_cleaned_with_readability2.xlsx")
df.head(3)

## Select relevant columns

df = df[["ID", "Newspaper", "Length", "Article", "modality_dummy", "reach_dummy", "clean text", "readabilityscore_text"]]
len(df)

## Descriptives

print("Overall sample")
print(df["readabilityscore_text"].mean())
print(df["readabilityscore_text"].std())

print("Print sample")
print(df[df["modality_dummy"] == 1]["readabilityscore_text"].mean())
print(df[df["modality_dummy"] == 1]["readabilityscore_text"].std())
print("")
print("Online sample")
print(df[df["modality_dummy"] == 0]["readabilityscore_text"].mean())
print(df[df["modality_dummy"] == 0]["readabilityscore_text"].std())

print("National sample")
print(df[df["reach_dummy"] == 1]["readabilityscore_text"].mean())
print(df[df["reach_dummy"] == 1]["readabilityscore_text"].std())
print("")
print("Regional sample")
print(df[df["reach_dummy"] == 0]["readabilityscore_text"].mean())
print(df[df["reach_dummy"] == 0]["readabilityscore_text"].std())

## Hypothesis test

print("RQ3a: Reach")
print(mannwhitneyu(df[df["reach_dummy"] == 1]["readabilityscore_text"], df[df["reach_dummy"] == 0]["readabilityscore_text"]))

print("RQ3b: Moality")
print(mannwhitneyu(df[df["modality_dummy"] == 1]["readabilityscore_text"], df[df["modality_dummy"] == 0]["readabilityscore_text"]))

## Comparisons

#grouping by outlet
df.groupby("Newspaper").mean()

#grouping by outlet
df.groupby("Newspaper").std()

from statsmodels.formula.api import ols

model = ols("readabilityscore_text ~ modality_dummy + reach_dummy + 1", df).fit()
print(model.summary()) 

print("AZ")
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["readabilityscore_text"].mean())
print(df[(df["Newspaper"]== "Aachener Zeitung") | (df["Newspaper"]== "aachener zeitung (www)")]["readabilityscore_text"].std())
print("RP")
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["readabilityscore_text"].mean())
print(df[(df["Newspaper"]== "Rheinische Post") | (df["Newspaper"]== "rheinische post (www)")]["readabilityscore_text"].std())
print("STZ")
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["readabilityscore_text"].mean())
print(df[(df["Newspaper"]== "Stuttgarter Zeitung") | (df["Newspaper"]== "stuttgarter zeitung (www)")]["readabilityscore_text"].std())
print("Welt")
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["readabilityscore_text"].mean())
print(df[(df["Newspaper"]== "Die Welt") | (df["Newspaper"]== "die welt (www)")]["readabilityscore_text"].std())
print("TS")
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["readabilityscore_text"].mean())
print(df[(df["Newspaper"]== "Der Tagesspiegel") | (df["Newspaper"]== "der tagesspiegel (www)")]["readabilityscore_text"].std())
print("SZ")
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["readabilityscore_text"].mean())
print(df[(df["Newspaper"]== "Süddeutsche Zeitung (inkl. Regionalausgaben)") | (df["Newspaper"]== "sueddeutschet politik (www)")]["readabilityscore_text"].std())
